meson: Update to latest git, finish support for ARM, general cleanups
* Fix build of opus_projection test

Needs the projection API to actually be built, and it needs to link
directly to the objects. This is also what the Makefile does.

At the same time, use kwargs: to set the test timeout so we don't need
to hard-code the default value.

Both these require a meson version bump.

* Set the libversion and darwin_versions to match the LT_VERSION

Now our binaries have the same ABI / soversion as the Autotools build.

* Reorganize intrinsics and rtcd support

* Add summary() support

* Increase warning_level instead of passing -Wextra

* Add missing stdint.h check